requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
openai
PyPDF2
python-docx
python-pptx
